Add sort optimization with after from Lucene
Lucene 8.7 introduces numeric sort optimization directly in comparators.
This means we don't need to have it in ES.
This removes the sort optimization code in ES, and
the only thing that is needed is setCanUsePoints on sortField.

This also will introduce sort optimization with search_after,
as Lucene directly suports sort optimization with search_after.