Introduce Vector Search Merge policy
For HNSW graph merge it is preferrable to merge the largest segment
with the smallest segments so we don't to recalculate the
largest graph.

This introduces the Vector Search merge policy that does it.
This policy can be enabled either by "index.merge.policy.type" : "vector_search"

```js
PUT index1
{
    "settings": {
        "index": {
            "merge.policy.type": "vector_search"
        }
    }
}
```

or by "index.mode" : "vector_search" setting:
```js
PUT index1
{
    "settings": {
        "index.mode": "vector_search"
    }
}
```